Add ability to track browser errors
A recent incident on Finder Frontend saw the filters on search pages
malfunctioning due to an error loading an asset file (which hadn't been
uploaded correctly). We could write a specific e2e test exercising the
filters, but ideally this functionality should already be tested
extensively in the app's own integration/system tests.

Instead, this adds a helper to enhance a test suite by keeping track of
console errors, Javascript exceptions, and network issues, which should
be able to catch many sources of unexpected failures in our frontend
code.

- Add `trackBrowserErrors` helper library
- Use `trackBrowserErrors` in the Finder Frontend e2e tests to begin
  with (possibly with a view to other apps opting in to this behaviour
  too)