Add full e2e test from publishing to frontend
We currently lack any automated test that gives us confidence
that publishing is working end to end, i.e. that an update from
a publishing app becomes visible on the frontend. Adding such a
test was highlighted as a recent [incident action](https://docs.google.com/document/d/14Hi5AkI6nf0l7Ot6KhjHu53Yq81F2vuRjjyMFtJWLcc/edit).

This commit adds a test that:

- Visits an existing 'test' document in Whitehall (available in
  all environments)
- Saves an edit to its draft content
- Verifies that the updated content is visible on the draft
  preview of that content

This tests the 'save draft' aspects of the publishing platform,
which is a reasonably thorough verification that the connection
from publishing app, to Publishing API, to Content Store is
working. It doesn't test the 'publish' action, but it doesn't
really need to, since that would just test the same aspects of
the publishing platform as above, only with a different final
destination (live content store instead of draft content store).
It also raises questions about how much of this automated test
activity should be public-facing, which is a can of worms we can
just avoid by sticking to the draft stack.

Trello: https://trello.com/c/H4gmTEtf/3489-explore-adding-full-e2e-publishing-test