Change site search result test
Google can sometimes change the ranking of search results on staging/integration.

Therefore we should use a more reliable search result for this test, as the 'Sign In' link can sometimes not exist depending on how the results are ranked.

This commit changes it to look for the universal-credit landing page, which is a lot more likely to always exist when searching for "Universal Credit"