Remove breaking expectation for SP
Latest changes for the Design System release have introduced a change to the routing in Specialist Publisher, so that the "/" route redirects to "/finders" by default, which is the index page of all finders, rather than the first finder the user has access to. Because of this, no "Add another" can be found on the page.

Removing this to unblock pub-api deployments, as well as the release of the DS itself.

Will reintroduce the correct expectation after the merge.