Update Publisher e2e to be specific with its page text search
Previously we were checking for the existence of the text "publisher"
without considering where that text was. As a result when we added
a banner, the test would get confused and return multiple instances
in a way not supported by the previous check.

Check has been updated to look within the specific header class.