Remove link check from search test
The specific results returned from a search request are determined by
Google Vertex. Sometimes there are problems with their algorithms so
the results returned are not the ones that we expect. This causes the
tests to fail and blocks deployment.

We have other, more details evaluations that run to check the quality of
search results, so we don't need to rely on the e2e test for that,
instead we should check things we can control, such as the user journey
(homepage to search page) being as expected.