Fix finder frontend site search tests
The `getByText` action was failing because it matched to 21 results on
the page, when it only expects to match to one.

The `toHaveCSS` assertion was failing because the CSS being searched for
is nested in a `ul` that's in a `li` that's in `ul`, so three levels
down the list.

Instead we look for the first occurrence of the link object on the page.
There is potential for a  false positive if the link to Universal Credit
is added somewhere else on the page, e.g. the footer, but this feels
unlikely. The check follows the same pattern as the [people test]

This test was tested locally against integration and passed.

[people test]: https://github.com/alphagov/govuk-e2e-tests/blob/938ccef6d519c236a3dca81a1d2108c8cd092107/tests/finder-frontend.spec.js#L11