Update `collections` app test
Remove the check for the "Cabinet ministers" heading from the "ministers" page test.

This heading does not exist when the ministers page is updated, this causes the e2e test to fail and block deployments.

Checking for the presence of the "Ministers" heading should be enough and is consistent with other page tests for `collections.