Request assets via `assets.` subdomain, not `www.`
It looks like support for serving assets via `www.`[^1] ended a few
weeks ago and that's caused this single test to start failing. The
correct way to directly request assets is via `assets.`.

Note: I wonder if it's possibly a coincidence that the assets domains
look like publishing app domains in all the envs, though? We could set
an `ASSETS_DOMAIN` variable alongside the `PUBLIC_DOMAIN` etc. There is
already an `assetsDomain` set for all envs in the helm charts. I don't
have any background knowledge as to why the assets domain looks the way
it does or why that `assetsDomain` was considered to be worth adding to
the config for each env, though, so, for now I'm happy to use the
`publishingAppUrl` util like the draft assets test already does.

[^1]: FWIW, that support depended on a special prefix route in Content
Store and some redirect code in Government Frontend.