Wait for Licensing admin redirect after Login
The page currently takes over 30 seconds to load.
This introduced a waiting for navigation to complete (new URL after login).
It will wait up to 60 seconds (60000 milliseconds).

Ideally we would want to have pagination on that page but this app is
maintained by external contractors so I'm not sure about a feasibility of that.