Skip CBM tests in other applications
For whatever reason, these haven't been reliable in Integration. It
would be nice to fix that, but currently the pain of blocking
deployments is bad enough that I'd rather not have the tests running at
all.

They'll still run in CBM's CI (I think) because the parent describe has
that tag. Maybe we could run them in a cron job too?