Simplify Content Block Manager tests
This updates the Content Block Manager tests to be more streamlined.
Now, rather than adding an embed code to a document and checking it
shows, then updating the value, we get the embed code, ensure it is
in both of our test draft documents. We then update the value and
only then do we test that the value shows.

This gets around some of the issues that we have in the non-production
environments, where sometimes blocks states get out of sync with the
content store, and our initial tests fail. Editing the block first
ensures the state of our block and what is in the content store and
Publishing API is in sync.

I’ve also added some rudimentary retrying functionality in the
`verifyUpdatedRateVisible` function. This replaces the
`waitForUrlToBeAvailable` method where we were polling the draft
preview. This function will now check if the URL is available, and
the updated value is present on the page, and will revisit the URL and
try again until Playwright times out if EITHER the URL is unavailable
OR the value is not visible on the page.

This test is now much nippier and does all the things we want to test
in a much smaller (and hopefully less flaky!) way.